move not "core-grodecoder" dependencies to "dev" group
