+++ pyproject.toml
@@ -5,15 +5,10 @@ description = "Add your description here"
 readme = "README.md"
 requires-python = ">=3.13"
 dependencies = [
-    "beautifulsoup4>=4.13.3",
-    "biopython>=1.85",
     "click>=8.1.8",
     "loguru>=0.7.3",
     "mdanalysis>=2.9.0",
-    "playwright>=1.51.0",
     "pydantic>=2.11.3",
-    "requests>=2.32.3",
-    "streamlit>=1.44.0",
 ]
 
 [build-system]
@@ -30,6 +25,11 @@ dev = [
     "pytest>=8.3.5",
     "ruff>=0.11.2",
     "types-requests>=2.32.4.20250809",
+    "beautifulsoup4>=4.13.3",
+    "biopython>=1.85",
+    "playwright>=1.51.0",
+    "requests>=2.32.3",
+    "streamlit>=1.44.0",
 ]
 
 [tool.ruff]
