+++ src/grodecoder/databases/__init__.py
@@ -1,4 +1,3 @@
-from . import csml, mad
 from .api import (
     get_amino_acid_definitions,
     get_amino_acid_names,
@@ -25,8 +24,6 @@ from .models import (
 )
 
 __all__ = [
-    "csml",
-    "mad",
     "get_amino_acid_definitions",
     "get_amino_acid_names",
     "get_amino_acid_name_map",
